Fix tpm2_checkquote when run on a big endian system
Various PCR related structures read by tpm2_checkquote are written
in little endian format, which needs to be taken into account when
run on a big endian system. Although little endian is unusual for
TPM, it is better to leave these data structure in little endian
format to not disrupt existing systems, so we do not convert these
structures to big endian on the side where they are written.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>